requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==4.0
django-allauth==0.47.0
django-cors-headers==3.10.1
django-countries==7.2.1
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
phonenumberslite==8.12.40
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stripe==2.65.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
